+++ src/components/dashboard/SendQRSummary.js
@@ -6,13 +6,11 @@ import React, { useEffect, useState } from 'react'
 import userStorage, { type TransactionEvent } from '../../lib/gundb/UserStorage'
 import logger from '../../lib/logger/pino-logger'
 import { useDialog } from '../../lib/undux/utils/dialog'
-import { withStyles } from '../../lib/styles'
 import { useWrappedGoodWallet } from '../../lib/wallet/useWrappedWallet'
 import { BackButton, useScreenState } from '../appNavigation/stackNavigation'
 import { CustomButton, Section, Wrapper } from '../common'
 import SummaryTable from '../common/view/SummaryTable'
 import TopBar from '../common/view/TopBar'
-import normalize from '../../lib/utils/normalizeText'
 import Config from '../../config/config'
 import SurveySend from './SurveySend'
 import { SEND_TITLE } from './utils/sendReceiveFlow'
@@ -30,11 +28,10 @@ const log = logger.child({ from: 'SendQRSummary' })
  * @param {any} props.screenProps
  * @param {any} props.styles
  */
-const SendQRSummary = ({ screenProps, styles }: AmountProps) => {
+const SendQRSummary = ({ screenProps }: AmountProps) => {
   const [screenState] = useScreenState(screenProps)
   const goodWallet = useWrappedGoodWallet()
   const [showDialog] = useDialog()
-  const [dialogSurvey, setDialogSurvey] = useState(true)
   const [survey, setSurvey] = useState('other')
   const [loading, setLoading] = useState(false)
   const [isValid, setIsValid] = useState(screenState.isValid)
@@ -46,34 +43,11 @@ const SendQRSummary = ({ screenProps, styles }: AmountProps) => {
     setProfile(profile)
   }
 
-  const checkSurvey = value => {
-    setSurvey(value)
-    setDialogSurvey(true)
-  }
-
   useEffect(() => {
     if (to) {
       updateRecepientProfile()
     }
   }, [to])
-
-  useEffect(() => {
-    if (Config.isEToro && dialogSurvey) {
-      setDialogSurvey(false)
-      showDialog({
-        content: <SurveySend handleCheckSurvey={checkSurvey} />,
-        buttons: [
-          {
-            style: styles.OkButton,
-            text: 'Ok',
-            onPress: dismiss => {
-              dismiss()
-            },
-          },
-        ],
-      })
-    }
-  }, [dialogSurvey])
 
   const faceRecognition = () => {
     return screenProps.push('FRIntro', { from: 'SendQRSummary' })
@@ -175,19 +149,9 @@ const SendQRSummary = ({ screenProps, styles }: AmountProps) => {
           </Section.Stack>
         </Section.Row>
       </Section>
+      <SurveySend handleCheckSurvey={setSurvey} />
     </Wrapper>
   )
-}
-
-const mapStylesToProps = ({ theme }) => {
-  return {
-    OkButton: {
-      borderWidth: 1,
-      borderColor: theme.colors.primary,
-      width: 100,
-      fontSize: normalize(14),
-    },
-  }
 }
 
 SendQRSummary.navigationOptions = {
@@ -201,4 +165,4 @@ SendQRSummary.shouldNavigateToComponent = props => {
   return (!!screenState.amount && !!screenState.to) || screenState.from
 }
 
-export default withStyles(mapStylesToProps)(SendQRSummary)
+export default SendQRSummary
